+++ tests/app_feature/test_advanced_request.py
@@ -72,7 +72,7 @@ class TestAdvancedRequest:
         assert r.status_code == 200
         # assert headers
         assert data.get('headers', {}).get('Pytest') == 'pytest'
-        assert data.get('headers', {}).get('User-Agent') == 'TcEx App: Pytest - 1.0.0'
+        assert data.get('headers', {}).get('User-Agent') == 'TcEx/2.0.26, Pytest/1.0.0'
         # assert params
         assert data.get('args', {}).get('one') == '1'
         assert data.get('args', {}).get('two') == ''
@@ -161,7 +161,7 @@ class TestAdvancedRequest:
         assert r.status_code == 200
         # assert headers
         assert data.get('headers', {}).get('Pytest') == 'pytest'
-        assert data.get('headers', {}).get('User-Agent') == 'TcEx App: Pytest - 1.0.0'
+        assert data.get('headers', {}).get('User-Agent') == 'TcEx/2.0.26, Pytest/1.0.0'
         # assert params
         assert data.get('args', {}).get('one') == '1'
         assert data.get('args', {}).get('two') is None
@@ -201,7 +201,7 @@ class TestAdvancedRequest:
         assert data.get('data') == tcex.args.tc_adv_req_body
         # assert headers
         assert data.get('headers', {}).get('Pytest') == 'pytest'
-        assert data.get('headers', {}).get('User-Agent') == 'TcEx App: Pytest - 1.0.0'
+        assert data.get('headers', {}).get('User-Agent') == 'TcEx/2.0.26, Pytest/1.0.0'
         # assert params
         assert data.get('args', {}).get('one') == '1'
 
@@ -240,7 +240,7 @@ class TestAdvancedRequest:
         assert data.get('data') == tcex.args.tc_adv_req_body.decode()
         # assert headers
         assert data.get('headers', {}).get('Pytest') == 'pytest'
-        assert data.get('headers', {}).get('User-Agent') == 'TcEx App: Pytest - 1.0.0'
+        assert data.get('headers', {}).get('User-Agent') == 'TcEx/2.0.26, Pytest/1.0.0'
         # assert params
         assert data.get('args', {}).get('one') == '1'
 
@@ -280,6 +280,6 @@ class TestAdvancedRequest:
         assert data.get('form', {}).get('two') == '2'
         # assert headers
         assert data.get('headers', {}).get('Pytest') == 'pytest'
-        assert data.get('headers', {}).get('User-Agent') == 'TcEx App: Pytest - 1.0.0'
+        assert data.get('headers', {}).get('User-Agent') == 'TcEx/2.0.26, Pytest/1.0.0'
         # assert params
         assert data.get('args', {}).get('one') == '1'